+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+## 0.24.3 – Mar 2, 2020
+### Featured
+Stability fixes
+
+### New
+- `min_shard_gen_utime` and `max_shard_gen_utime` fields in `block.master`
+
+### Fixed
+- joined objects returned as `null` if joined object inserted in DB later than parent object.   
+
 ## 0.24.2 – Feb 19, 2020
 ### Featured
 Ability to set restrictions to accounts for particular access keys